+++ tests/Generator/Type/ClassTest.php
@@ -42,28 +42,8 @@ class Generator_Type_ClassTest extends Unittest_TestCase
 		$this->assertSame('TestCountable, TestSortable', $params['implements']);
 		$this->assertSame('TestClassParent', $params['extends']);
 
-		$this->assertCount(1, $params['methods']);
-		$this->assertCount(2, $params['methods']['public']);
-		$this->assertArrayNotHasKey('abstract', $params['methods']);
-
-		// Abstract classes should not implement a parent's methods
-		$this->assertArrayNotHasKey('public_method', $params['methods']['public']);
-		$this->assertArrayNotHasKey('abstract_method', $params['methods']['public']);
-
-		// Classes should always implement interfaces
-		$this->assertArrayNotHasKey('abstract_method', $params['methods']['public']);
-		$this->assertSame('TestCountable', $params['methods']['public']['count']['class']);
-		$this->assertNotEmpty($params['methods']['public']['count']['doccomment']);
-		$this->assertFalse($params['methods']['public']['count']['abstract']);
-		$this->assertArrayHasKey('body', $params['methods']['public']['count']);
-		$this->assertNotRegExp('/parent::/', $params['methods']['public']['count']['body']);
-
-		$this->assertArrayHasKey('sort', $params['methods']['public']);
-		$this->assertSame('TestSortable', $params['methods']['public']['sort']['class']);
-		$this->assertNotEmpty($params['methods']['public']['sort']['doccomment']);
-		$this->assertFalse($params['methods']['public']['sort']['abstract']);
-		$this->assertArrayHasKey('body', $params['methods']['public']['sort']);
-		$this->assertNotRegExp('/parent::/', $params['methods']['public']['sort']['body']);
+		// Abstract classes should not implement inherited abstract methods
+		$this->assertCount(0, $params['methods']);
 	}
 
 	/**
@@ -76,27 +56,48 @@ class Generator_Type_ClassTest extends Unittest_TestCase
 		$module = basename(dirname(dirname(dirname(dirname(__FILE__)))));
 		$type = new Generator_Type_Class('Foo');
 
-		$type->module($module)
-			->extend('TestClassParent')
+		$type->extend('TestClassParent')
+			->implement('TestCountable')
+			->implement('TestSortable')
+			->module($module)
 			->blank(FALSE);
 
 		$params = $type->params();
 		$this->assertSame($module, $type->module());
 		$this->assertArrayNotHasKey('abstract', $params);
 		$this->assertSame('TestClassParent', $params['extends']);
+		$this->assertContains('TestCountable', $params['implements']);
+		$this->assertContains('TestSortable', $params['implements']);
+		$this->assertArrayNotHasKey('methods', $params);
 
 		$type->render();
 		$params = $type->params();
 
-		// Concrete classes should implement any parent's abstract methods
+		$this->assertCount(1, $params['methods']);
+		$this->assertCount(3, $params['methods']['public']);
+		$this->assertArrayNotHasKey('abstract', $params['methods']);
 		$this->assertArrayNotHasKey('public_method', $params['methods']['public']);
-		$this->assertArrayHasKey('abstract_method', $params['methods']['public']);
 
-		$this->assertSame('TestClassParent', $params['methods']['public']['abstract_method']['class']);
+		// Concrete classes should implement a parent's abstract methods
+		$this->assertArrayHasKey('abstract_method', $params['methods']['public']);
 		$this->assertNotEmpty($params['methods']['public']['abstract_method']['doccomment']);
 		$this->assertFalse($params['methods']['public']['abstract_method']['abstract']);
 		$this->assertArrayHasKey('body', $params['methods']['public']['abstract_method']);
 		$this->assertNotRegExp('/parent::/', $params['methods']['public']['abstract_method']['body']);
+
+		// Concrete classes should implement all interface methods
+		$this->assertSame('TestCountable', $params['methods']['public']['count']['class']);
+		$this->assertNotEmpty($params['methods']['public']['count']['doccomment']);
+		$this->assertFalse($params['methods']['public']['count']['abstract']);
+		$this->assertArrayHasKey('body', $params['methods']['public']['count']);
+		$this->assertNotRegExp('/parent::/', $params['methods']['public']['count']['body']);
+
+		$this->assertArrayHasKey('sort', $params['methods']['public']);
+		$this->assertSame('TestSortable', $params['methods']['public']['sort']['class']);
+		$this->assertNotEmpty($params['methods']['public']['sort']['doccomment']);
+		$this->assertFalse($params['methods']['public']['sort']['abstract']);
+		$this->assertArrayHasKey('body', $params['methods']['public']['sort']);
+		$this->assertNotRegExp('/parent::/', $params['methods']['public']['sort']['body']);
 	}
 
 } // End Generator_Type_ClassTest